requirements.txt
google-api-python-client==1.4.1
httplib2==0.9.1
oauth2client==1.4.12
pyasn1-modules==0.0.7
pyasn1==0.1.8
rsa==3.2
simplejson==3.8.0
six==1.9
uritemplate==0.6
functools32==3.2.3.post2; python_version < "3"
